Fix alibi slopes dtype for mixed-precision and flash attention
Slopes must necessarily be in float32 for flash attention

+++ bertblocks/modeling/attention.py
@@ -1,3 +1,4 @@
+import torch
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
@@ -123,7 +124,7 @@ class Attention(nn.Module):
                 max_seq_len,
                 self.num_heads,
                 self.head_dim,
-                alibi_slopes=AlibiPositionalEncoding.get_slopes(self.num_heads, device=qkv.device, dtype=qkv.dtype),
+                alibi_slopes=AlibiPositionalEncoding.get_slopes(self.num_heads, device=qkv.device, dtype=torch.float32),
                 local_attention=self.local_attention,
                 dropout_p=self.dropout_p if self.training else 0.0,
                 deterministic=self.deterministic,
